requirements.txt
munkres==1.1.4
numpy==2.2.6
opencv-contrib-python==4.12.0.88
pygame==2.6.1
pygame_gui==0.6.14
PyQt6==6.10.1
PyQt6-Qt6==6.10.1
PyQt6_sip==13.10.3
pyserial==3.5
python-i18n==0.3.9
pyusb==1.3.1
pyzmq==27.1.0
setuptools==80.9.0
wheel==0.45.1